some refactoring to make the external function more concise
