+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Tue Jul 09 14:13:52 CEST 2019
-gradle.version=5.3.1
+#Sat Jul 13 12:09:24 CEST 2019
+gradle.version=5.4.1
