+++ src/main/kotlin/day1/homeExercise/BirthdayKata.kt
@@ -1,32 +1,37 @@
 package day1.homeExercise
 
-import day1.example3.flatMap
-import day1.example3.sequence
-import day1.exercise1.andThen
+import day1.example3.Outcome
+import java.time.LocalDate
 
 //see http://matteo.vaccari.name/blog/archives/154 for the Kata rules
+
+class SendGreetingsToAll(
+    private val loadEmployees: (FileName) -> Outcome<ProgramError, List<Employee>>,
+    private val employeeBorneToday: (Employee) -> Boolean,
+    private val sendBirthDayGreetingMail: (Employee) -> Outcome<MailSendingFailure, Unit>
+) : (FileName) -> Unit {
+
+    override fun invoke(sourceFile: FileName) {
+        loadEmployees(sourceFile)
+            .map {employees ->
+                employees
+                    .filter(employeeBorneToday)
+                    .map(sendBirthDayGreetingMail)
+            }
+    }
+
+}
+
+data class Employee(
+    val lastName: String,
+    val firstName: String,
+    val birthDate: LocalDate,
+    val emailAddress: EmailAddress
+)
+
+data class EmailAddress(val value: String)
 
 sealed class ProgramError
 data class FileNotFound(val path: String) : ProgramError()
 data class ParseError(val source: String) : ProgramError()
 data class MailSendingFailure(val emailMessage: EmailMessage) : ProgramError()
-
-class SendGreetingsToAll(
-    private val readCsv: ReadCsv,
-    private val parseEmployee: ParseEmployee,
-    private val employeeBorneToday: EmployeeBirthdayFilter,
-    private val mailCompose: ComposeBirthdayEmailMessage,
-    private val sendEmail: SendEmail
-) : (FileName) -> Unit {
-    override fun invoke(sourceFile: FileName) {
-        readCsv(sourceFile)
-            .flatMap { file -> file.rows.map(parseEmployee).sequence() }
-            .map { employees ->
-                employees
-                    .filter(employeeBorneToday)
-                    .forEach { employee ->
-                        (mailCompose andThen sendEmail)(employee)
-                    }
-            }
-    }
-}
