--- src/main/kotlin/day1/homeExercise/Domain.kt
@@ -1,12 +1,0 @@
-package day1.homeExercise
-
-import java.time.LocalDate
-
-data class Employee(
-    val lastName: String,
-    val firstName: String,
-    val birthDate: LocalDate,
-    val emailAddress: EmailAddress
-)
-
-data class EmailAddress(val value: String)
